pyproject.toml
[tool.poetry]
name = "duckdb-docker-etl"
version = "0.1.0"
description = ""
authors = ["henriquehashimoto <henriquehashimoto518@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gdown = "^5.2.0"
duckdb = "^1.0.0"
streamlit = "^1.37.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.32"
fsspec = "^2024.6.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
